+++ include/stlsoft/filesystem/io_functions/write_string.hpp
@@ -4,7 +4,7 @@
  * Purpose: Definition of stlsoft::write_string<>() function template.
  *
  * Created: 22nd February 2024
- * Updated: 11th March 2024
+ * Updated: 5th May 2024
  *
  * Home:    http://stlsoft.org/
  *
@@ -53,8 +53,8 @@
 #ifndef STLSOFT_DOCUMENTATION_SKIP_SECTION
 # define STLSOFT_VER_STLSOFT_FILESYSTEM_IO_FUNCTIONS_HPP_WRITE_STRING_MAJOR     1
 # define STLSOFT_VER_STLSOFT_FILESYSTEM_IO_FUNCTIONS_HPP_WRITE_STRING_MINOR     0
-# define STLSOFT_VER_STLSOFT_FILESYSTEM_IO_FUNCTIONS_HPP_WRITE_STRING_REVISION  1
-# define STLSOFT_VER_STLSOFT_FILESYSTEM_IO_FUNCTIONS_HPP_WRITE_STRING_EDIT      2
+# define STLSOFT_VER_STLSOFT_FILESYSTEM_IO_FUNCTIONS_HPP_WRITE_STRING_REVISION  2
+# define STLSOFT_VER_STLSOFT_FILESYSTEM_IO_FUNCTIONS_HPP_WRITE_STRING_EDIT      3
 #endif /* !STLSOFT_DOCUMENTATION_SKIP_SECTION */
 
 
@@ -108,7 +108,7 @@ namespace stlsoft
 
 namespace write_string_impl
 {
-    int
+    ss_size_t
     write_string_3_(
         FILE*               stm
     ,   ss_char_a_t const*  s
@@ -147,7 +147,7 @@ template <
     ss_typename_param_k T_stream
 ,   ss_typename_param_k T_string
 >
-int
+ss_size_t
 write_string(
     T_stream&       stm
 ,   T_string const& s
@@ -176,7 +176,7 @@ template <
     ss_typename_param_k T_stream
 ,   ss_typename_param_k T_char
 >
-int
+ss_size_t
 write_string(
     T_stream&       stm
 ,   T_char const*   s