+++ include/winstl/system/temporary_file_name.hpp
@@ -1,15 +1,15 @@
 /* /////////////////////////////////////////////////////////////////////////
- * File:        winstl/system/temporary_file_name.hpp
+ * File:    winstl/system/temporary_file_name.hpp
  *
- * Purpose:     Simple class that gets, and makes accessible, a temporary
- *              file name.
+ * Purpose: Simple class that gets, and makes accessible, a temporary file
+ *          name.
  *
- * Created:     5th June 2011
- * Updated:     11th March 2024
+ * Created: 5th June 2011
+ * Updated: 5th May 2024
  *
- * Thanks to:   Pablo Aguilar for requesting this component.
+ * Thanks:  Pablo Aguilar for requesting this component.
  *
- * Home:        http://stlsoft.org/
+ * Home:    http://stlsoft.org/
  *
  * Copyright (c) 2019-2024, Matthew Wilson and Synesis Information Systems
  * Copyright (c) 2011-2019, Matthew Wilson and Synesis Software
@@ -57,8 +57,8 @@
 #ifndef STLSOFT_DOCUMENTATION_SKIP_SECTION
 # define WINSTL_VER_WINSTL_SYSTEM_HPP_TEMPORARY_FILE_NAME_MAJOR     1
 # define WINSTL_VER_WINSTL_SYSTEM_HPP_TEMPORARY_FILE_NAME_MINOR     2
-# define WINSTL_VER_WINSTL_SYSTEM_HPP_TEMPORARY_FILE_NAME_REVISION  2
-# define WINSTL_VER_WINSTL_SYSTEM_HPP_TEMPORARY_FILE_NAME_EDIT      23
+# define WINSTL_VER_WINSTL_SYSTEM_HPP_TEMPORARY_FILE_NAME_REVISION  4
+# define WINSTL_VER_WINSTL_SYSTEM_HPP_TEMPORARY_FILE_NAME_EDIT      25
 #endif /* !STLSOFT_DOCUMENTATION_SKIP_SECTION */
 
 
@@ -198,7 +198,7 @@ private:
             DWORD const e = WINSTL_API_EXTERNAL_ErrorHandling_GetLastError();
 
 #ifdef STLSOFT_CF_EXCEPTION_SUPPORT
-            STLSOFT_THROW_X(windows_exception("could not elicit temporary directory", e));
+            STLSOFT_THROW_X(winstl_exception("could not elicit temporary directory", e));
 #else /* ? STLSOFT_CF_EXCEPTION_SUPPORT */
 
             if (0 != cchBuff)
@@ -210,14 +210,14 @@ private:
 #endif /* STLSOFT_CF_EXCEPTION_SUPPORT */
         }
 
-        UINT const  r   =   ::GetTempFileNameA(&dirPath_[0], "wst", 0, &filePath_[0]);
+        UINT const r = ::GetTempFileNameA(&dirPath_[0], "wst", 0, &filePath_[0]);
 
         if (0 == r)
         {
             DWORD const e = WINSTL_API_EXTERNAL_ErrorHandling_GetLastError();
 
 #ifdef STLSOFT_CF_EXCEPTION_SUPPORT
-            STLSOFT_THROW_X(windows_exception("could not elicit temporary file name", e));
+            STLSOFT_THROW_X(winstl_exception("could not elicit temporary file name", e));
 #else /* ? STLSOFT_CF_EXCEPTION_SUPPORT */
 
             if (0 != cchBuff)
@@ -229,7 +229,7 @@ private:
 #endif /* STLSOFT_CF_EXCEPTION_SUPPORT */
         }
 
-        size_t      cch =   traits_type_::str_len(filePath_);
+        size_t cch = ::str_len(filePath_);
 
         if (0 != cchBuff)
         {
@@ -256,14 +256,14 @@ private:
         ws_char_w_t dirPath_[1 + MAX_PATH] = { 0 } ;
         ws_char_w_t filePath_[1 + MAX_PATH] = { 0 } ;
 
-        DWORD const dw  =   ::GetTempPathW(STLSOFT_NUM_ELEMENTS(dirPath_), &dirPath_[0]);
+        DWORD const dw = ::GetTempPathW(STLSOFT_NUM_ELEMENTS(dirPath_), &dirPath_[0]);
 
         if (0 == dw)
         {
             DWORD const e = WINSTL_API_EXTERNAL_ErrorHandling_GetLastError();
 
 #ifdef STLSOFT_CF_EXCEPTION_SUPPORT
-            STLSOFT_THROW_X(windows_exception("could not elicit temporary directory", e));
+            STLSOFT_THROW_X(winstl_exception("could not elicit temporary directory", e));
 #else /* ? STLSOFT_CF_EXCEPTION_SUPPORT */
 
             if (0 != cchBuff)
@@ -275,14 +275,14 @@ private:
 #endif /* STLSOFT_CF_EXCEPTION_SUPPORT */
         }
 
-        UINT const  r   =   ::GetTempFileNameW(&dirPath_[0], L"wst", 0, &filePath_[0]);
+        UINT const r = ::GetTempFileNameW(&dirPath_[0], L"wst", 0, &filePath_[0]);
 
         if (0 == r)
         {
             DWORD const e = WINSTL_API_EXTERNAL_ErrorHandling_GetLastError();
 
 #ifdef STLSOFT_CF_EXCEPTION_SUPPORT
-            STLSOFT_THROW_X(windows_exception("could not elicit temporary file name", e));
+            STLSOFT_THROW_X(winstl_exception("could not elicit temporary file name", e));
 #else /* ? STLSOFT_CF_EXCEPTION_SUPPORT */
 
             if (0 != cchBuff)
@@ -294,7 +294,7 @@ private:
 #endif /* STLSOFT_CF_EXCEPTION_SUPPORT */
         }
 
-        size_t      cch =   traits_type_::str_len(filePath_);
+        size_t cch = traits_type_::str_len(filePath_);
 
         if (0 != cchBuff)
         {