+++ test/unit/stlsoft/filesystem/test.unit.stlsoft.filesystem.io_functions.read_line/entry.cpp
@@ -4,7 +4,7 @@
  * Purpose: Unit-test
  *
  * Created: 11th August 2008
- * Updated: 22nd February 2024
+ * Updated: 5th May 2024
  *
  * ////////////////////////////////////////////////////////////////////// */
 
@@ -366,7 +366,7 @@ int main(int argc, char **argv)
 
     XTESTS_COMMANDLINE_PARSEVERBOSITY(argc, argv, &verbosity);
 
-    if(XTESTS_START_RUNNER("test.unit.stlsoft.filesystem.io_functions.read_line", verbosity))
+    if (XTESTS_START_RUNNER("test.unit.stlsoft.filesystem.io_functions.read_line", verbosity))
     {
         XTESTS_RUN_CASE(test_1_0);
         XTESTS_RUN_CASE(test_1_1);