+++ test/unit/stlsoft/string/test.unit.stlsoft.string.string_slice/entry.cpp
@@ -505,7 +505,7 @@ static void test_insertion_4(void)
 
 
 #if __cplusplus >= 201402L
-    std::string expected = ([&s2, &s3]() {
+    std::string expected = ([&s2, &s3, FIELD_WIDTH]() {
 #else
     struct Expected
     {